+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Next Release
 
 * [#PR_NUMBER](https://github.com/jouba/pull/PR_NUMBER): DESCRIPTION. - [@CONTRIBUTOR](https://github.com/CONTRIBUTOR).
-
+* [#5](https://github.com/jouba/pull/5): Dont splat events params. - [@diegodurs](https://github.com/diegodurs).
 * [#6](https://github.com/jouba/pull/6): Add changelog and contributing files. - [@gregory](https://github.com/gregory).
 * [#8](https://github.com/jouba/pull/8): [fix ruby <2 to_h -> to_hash](https://github.com/gregory/jouba/issues/7) . - [@gregory](https://github.com/gregory).
 * [#1](https://github.com/gregory/jouba/pull/1): Fix README, remove unnecessary include in spec. - [@diegodurs](https://github.com/diegodurs).